flask/xichuangzhu-master/requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
flask-debugtoolbar
Flask-Uploads
Flask-Mail
Flask-Script
Flask-Migrate

pypinyin
#OpenCC
permission
requests
markdown2
#MySQL-python
gunicorn
#alioss
#lxmlb1
cssselect
six
livereload
glob2

#fabric
blinker
raven

nose
coverage